Add a new way of doing part 1
Missy had a bug, so I was helping her debug, and in the process, came up with a merged way of her idea and what I thought she was doing. By using ranges, we can just calculate the points by finding the max of the maxes and min of the mins. Apparently... the fact that there is a beacon already on the line doesn't matter. Which is confusing to me, and probably means that I am actually missing something/there is a bug (ie, I have an off by one bug, and I guess that's ok, though I'd really like to understand WHY I have an off by one bug). It is also O(n) on the list of signals. So basically instantaneous, since you don't have to iterate through all the points like I was originally doing.